requirements.txt
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-allauth==0.61.1
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
requests==2.31.0
python-dateutil==2.8.2 